Add support for window stack references (%1, %@, etc) to the `exec` command
Window references will be replaced on the command-line in the same
position that they occurred. In case of %@ (all windows), multiple
arguments are inserted.

For example, if the window list is two entries: 123, 456

`exec echo Hello %2` will run the command exactly as this list
  [echo, "Hello", "456"]

`exec echo World %@` will run the command exactly as this list
  [echo, "World", "123", "456"]
with each window id being a separate argument.

For #431